# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "A comprehensive news aggregator API that fetches articles from NewsAPI, The Guardian, and The New York Times. Provides advanced search, filtering, and personalization features.",
        "contact": {
            "name": "API Support",
            "email": "support@newsaggregator.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/sources/{id}": {
            "get": {
                "tags": [
                    "Sources"
                ],
                "summary": "Get a single news source",
                "description": "Retrieve detailed information about a specific news source",
                "operationId": "9595ed26f1cff3f6c4d85b0e5e9ad5a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Source ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Source information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "key": {
                                            "type": "string",
                                            "example": "newsapi"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "NewsAPI"
                                        },
                                        "base_url": {
                                            "type": "string",
                                            "example": "https://newsapi.org/v2"
                                        },
                                        "articles_count": {
                                            "type": "integer",
                                            "example": 1250
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Source not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Source not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/preferences": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get user preferences",
                "description": "Retrieve the authenticated user's saved preferences for sources, categories, and authors",
                "operationId": "9d5a4e9bfa000aaafb6b6dc5ff600591",
                "responses": {
                    "200": {
                        "description": "User preferences",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "preferences": {
                                            "properties": {
                                                "sources": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "newsapi"
                                                    }
                                                },
                                                "categories": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Technology"
                                                    }
                                                },
                                                "authors": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Jane Smith"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Update user preferences",
                "description": "Update the authenticated user's preferences for personalized article feed",
                "operationId": "d37cd7270c032aa280ae454a296aeeca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "preferences"
                                ],
                                "properties": {
                                    "preferences": {
                                        "properties": {
                                            "sources": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "enum": [
                                                        "newsapi",
                                                        "guardian",
                                                        "nytimes"
                                                    ]
                                                },
                                                "example": [
                                                    "newsapi",
                                                    "guardian"
                                                ]
                                            },
                                            "categories": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "Technology",
                                                    "Business"
                                                ]
                                            },
                                            "authors": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "Jane Smith"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preferences updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preferences updated successfully"
                                        },
                                        "preference": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Reset user preferences",
                "description": "Delete the authenticated user's preferences and reset to defaults",
                "operationId": "ca41e40bad2d78a6ac0ddc3020a64ebe",
                "responses": {
                    "200": {
                        "description": "Preferences reset successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preferences reset successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}